feat: adds opt in button slot to sidebar layout
