+++ src/layouts/sidebar-sidecar/index.tsx
@@ -38,6 +38,7 @@ const SidebarSidecarLayoutContent = ({
   githubFileUrl,
   headings,
   openConsentManager,
+  optOutButtonSlot,
   sidebarSlot,
   sidecarSlot,
   sidebarNavDataLevels,
@@ -110,7 +111,10 @@ const SidebarSidecarLayoutContent = ({
         <div className={s.mainArea}>
           <div className={s.main}>
             <main id="main">
-              {breadcrumbLinks && <BreadcrumbBar links={breadcrumbLinks} />}
+              <span className={s.breadcrumbOptOutGroup}>
+                {breadcrumbLinks && <BreadcrumbBar links={breadcrumbLinks} />}
+                {optOutButtonSlot && optOutButtonSlot}
+              </span>
               {children}
               {githubFileUrl && (
                 <EditOnGithubLink
