+++ src/layouts/sidebar-sidecar/sidebar-sidecar-layout.module.css
@@ -88,3 +88,9 @@ for further details.
   top: calc(var(--sticky-bars-height) + var(--main-area-top-padding));
   width: 220px;
 }
+
+.breadcrumbOptOutGroup {
+  display: flex;
+  justify-content: space-between;
+  align-items: baseline;
+}
