+++ src/layouts/sidebar-sidecar/types.ts
@@ -13,6 +13,7 @@ interface BaseProps {
   githubFileUrl?: string
   openConsentManager?: () => void
   sidebarNavDataLevels?: SidebarProps[]
+  optOutButtonSlot?: ReactElement
 }
 
 /**
