+++ src/views/product-tutorials-view/index.tsx
@@ -4,14 +4,12 @@ import CoreDevDotLayout from 'layouts/core-dev-dot-layout'
 import ProductCollectionsSidebar, {
   ProductCollectionsSidebarProps,
 } from 'components/tutorials-sidebar/compositions/product-collections-sidebar'
-import Button from 'components/button'
-import { IconExternalLink16 } from '@hashicorp/flight-icons/svg-react/external-link-16'
 import { ProductTutorialsSitemap } from './components'
 import { ProductTutorialsViewProps } from './server'
 import ProductViewContent from './components/product-view-content'
+import { OptOutButton } from './components/opt-out-button'
 import { getOverviewHeading } from './helpers/heading-helpers'
 import s from './product-tutorials-view.module.css'
-import { useRouter } from 'next/router'
 
 function ProductTutorialsView({
   layoutProps,
@@ -51,9 +49,9 @@ function ProductTutorialsView({
           sections={layoutProps.sidebarSections}
         />
       }
+      optOutButtonSlot={<OptOutButton />}
     >
       <PageHeading />
-      <OptOutButton />
       <ProductViewContent
         blocks={blocks}
         inlineCollections={inlineCollections}
@@ -73,22 +71,3 @@ function ProductTutorialsView({
 
 ProductTutorialsView.layout = CoreDevDotLayout
 export default ProductTutorialsView
-
-const LEARN_BASE_URL = 'https://learn.hashicorp.com/'
-
-function OptOutButton() {
-  const router = useRouter()
-  // handle the remapping, we have all the data needed here to construct the right path
-  // currently this will only work for product tutorial landing pages
-  const learnPath = router.pathname.split('/')[0]
-  const url = new URL(learnPath + '?betaOptOut=true', LEARN_BASE_URL)
-  return (
-    <Button
-      color="tertiary"
-      text="Opt out"
-      icon={<IconExternalLink16 />}
-      iconPosition="trailing"
-      onClick={() => window.location.assign(url)}
-    />
-  )
-}
